- restructuring movieDetail
