+++ TheMovieDb/Modules/MovieDetail/MovieDetailRemoteDataManager.swift
@@ -13,7 +13,11 @@ class MovieDetailRemoteDataManager: MovieDetailRemoteDataManagerInputProtocol {
     private let group = DispatchGroup()
     private let defaultParameters = APIParameters()
     var movies: [MovieDetailSections: [Movie]] = [:]
-    var service: APIMoviesProtocol?
+    private let service: APIMoviesProtocol
+    
+    init(service: APIMoviesProtocol) {
+        self.service = service
+    }
     
     func fetchRelatedMovies() {
         MovieDetailSections.allCases.forEach { fetchData(typeMovieSection: $0) }
@@ -24,7 +28,7 @@ class MovieDetailRemoteDataManager: MovieDetailRemoteDataManagerInputProtocol {
     
     private func fetchData(typeMovieSection: MovieDetailSections) {
         group.enter()
-        service?.fetchData(endPoint: typeMovieSection.path, with: defaultParameters, completion: { [weak self] (response: Result<Movies, Error>) in
+        service.fetchData(endPoint: typeMovieSection.path, with: defaultParameters, completion: { [weak self] (response: Result<Movies, Error>) in
             switch response {
             case .failure(let error):
                 debugPrint(error)
