+++ TheMovieDb/Modules/Searching/SearchingRouter.swift
@@ -12,7 +12,7 @@ import UIKit
 final class SearchingRouter: SearchingRouterProtocol {
     
     func showMovieDetail(from view: SearchingViewProtocol, with movie: Movie) {
-        let controller = MovieDetailWireFrame.createMovieDetailModule(with: movie)
+        let controller = MovieDetailBuilder.createModule(with: movie)
         guard let view = view as? SearchingView else { return }
         view.navigationController?.pushViewController(controller, animated: true)
     }
